Fix unit test.
